+++ src/calculatricegraphique/DIV.java
@@ -16,7 +16,7 @@ public class DIV extends Binary {
         right = e2;
     }
 
-    public int eval() {
+    public double eval() {
         return left.eval() / right.eval();
     }
 }
